project/requirements.txt
streamlit==1.38.0
tensorflow==2.17.0
numpy==1.26.0
Pillow==10.0.0